+++ packages/fiori-annotation-api/src/xml/writer.ts
@@ -37,7 +37,7 @@ import type {
     UpdateElementName,
     XMLDocumentChange
 } from './changes';
-import { REPLACE_ELEMENT_CONTENT } from './changes';
+import { DELETE_ELEMENT, REPLACE_ELEMENT_CONTENT } from './changes';
 import { getNodeFromPointer } from './pointer';
 
 const printOptions: typeof defaultPrintOptions = { ...defaultPrintOptions, useSnippetSyntax: false };
@@ -79,7 +79,7 @@ export class XMLWriter {
         const changes = preprocessChanges(this.changes, this.document);
         const batches = getBatches(changes);
         for (const pointer of Object.keys(batches)) {
-            edits.push(...this.getTextEditsForPointer(pointer, batches[pointer]));
+            edits.push(...this.getTextEditsForPointer(pointer, batches[pointer], changes));
         }
         edits.sort(compareByRange);
         return edits;
@@ -102,10 +102,12 @@ export class XMLWriter {
      *
      * @param pointer
      * @param byType
+     * @param changes
      */
     private getTextEditsForPointer(
         pointer: string,
-        byType: Map<XMLDocumentChange['type'], XMLDocumentChange[]>
+        byType: Map<XMLDocumentChange['type'], XMLDocumentChange[]>,
+        changes: XMLDocumentChange[]
     ): TextEdit[] {
         const edits: TextEdit[] = [];
         const element = getNodeFromPointer(this.document, pointer);
@@ -132,7 +134,9 @@ export class XMLWriter {
                     attributeInserts: (byType.get('insert-attribute') ?? []) as InsertAttribute[],
                     moveInCollection: (byType.get('move-collection-value') ?? []) as MoveCollectionValue[]
                 };
-                edits.push(...this.handleXmlElementChanges(elementChanges, element, pointer, childIndentLevel));
+                edits.push(
+                    ...this.handleXmlElementChanges(elementChanges, element, pointer, childIndentLevel, changes)
+                );
                 break;
             }
             case 'XMLAttribute': {
@@ -161,6 +165,7 @@ export class XMLWriter {
      * @param element
      * @param pointer
      * @param childIndentLevel
+     * @param changes
      */
     private handleXmlElementChanges(
         elementChanges: {
@@ -172,7 +177,8 @@ export class XMLWriter {
         },
         element: XMLElement,
         pointer: string,
-        childIndentLevel: number
+        childIndentLevel: number,
+        changes: XMLDocumentChange[]
     ) {
         const edits: TextEdit[] = [];
         const { replacements, contentReplacements, elementDeletions, attributeInserts, moveInCollection } =
@@ -191,7 +197,7 @@ export class XMLWriter {
                 if (!insertPosition) {
                     continue;
                 }
-                const { textEdits, text } = this.prepareXmlElementMoveChange(moveChange);
+                const { textEdits, text } = this.prepareXmlElementMoveChange(moveChange, changes);
                 edits.push(...textEdits);
                 edits.push(...handleXmlElementMoveChange(element, childIndentLevel, text, insertPosition));
             }
@@ -204,8 +210,12 @@ export class XMLWriter {
     /**
      *
      * @param moveChange
+     * @param changes
      */
-    private prepareXmlElementMoveChange(moveChange: MoveCollectionValue): {
+    private prepareXmlElementMoveChange(
+        moveChange: MoveCollectionValue,
+        changes: XMLDocumentChange[]
+    ): {
         textEdits: TextEdit[];
         text: string[];
     } {
@@ -219,6 +229,18 @@ export class XMLWriter {
                 continue;
             }
             text.push(this.textDocument.getText(range));
+            if (
+                changes.some((change) => {
+                    if (change.type !== DELETE_ELEMENT) {
+                        return false;
+                    }
+                    const element = getNodeFromPointer(this.document, change.pointer);
+                    const changeRange = element?.type === 'XMLElement' && transformRange(element.position);
+                    return changeRange && rangeContained(changeRange, range);
+                })
+            ) {
+                continue;
+            }
             textEdits.push(TextEdit.del(range));
         }
         return { textEdits, text };