+++ packages/odata-service-inquirer/test/unit/prompts/sap-system/abap-on-prem/questions.test.ts
@@ -108,6 +108,18 @@ describe('questions', () => {
                 "when": [Function],
               },
               {
+                "additionalMessages": [Function],
+                "default": false,
+                "guiOptions": {
+                  "breadcrumb": "Store Credentials",
+                },
+                "message": "Do you want to store the system credentials?",
+                "name": "abapOnPrem:storeSystemCredentials",
+                "type": "confirm",
+                "validate": [Function],
+                "when": [Function],
+              },
+              {
                 "default": [Function],
                 "guiOptions": {
                   "applyDefaultWhenDirty": true,