+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.4.7
+
+### Patch Changes
+
+-   Updated dependencies [afb4be5]
+    -   @sap-ux/ui-components@1.26.15
+    -   @sap-ux/ui-prompting@0.5.6
+
+## 0.4.6
+
+### Patch Changes
+
+-   Updated dependencies [cfe9c13]
+    -   @sap-ux/ui-components@1.26.14
+    -   @sap-ux/ui-prompting@0.5.5
+
 ## 0.4.5
 
 ### Patch Changes