+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,78 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.5.0
+
+### Minor Changes
+
+-   aa2c7df: Bump specVersion
+
+### Patch Changes
+
+-   @sap-ux/odata-service-inquirer@2.11.4
+
+## 0.4.46
+
+### Patch Changes
+
+-   Updated dependencies [2e75589]
+    -   @sap-ux/odata-service-inquirer@2.11.4
+
+## 0.4.45
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.9.3
+-   @sap-ux/cf-deploy-config-sub-generator@0.2.80
+-   @sap-ux/deploy-config-generator-shared@0.1.52
+-   @sap-ux/odata-service-inquirer@2.11.3
+-   @sap-ux/abap-deploy-config-sub-generator@0.1.95
+
+## 0.4.44
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.13.33
+-   @sap-ux/inquirer-common@0.9.2
+-   @sap-ux/odata-service-inquirer@2.11.2
+-   @sap-ux/abap-deploy-config-sub-generator@0.1.94
+-   @sap-ux/cf-deploy-config-sub-generator@0.2.79
+-   @sap-ux/deploy-config-generator-shared@0.1.52
+
+## 0.4.43
+
+### Patch Changes
+
+-   cfe9c13: Add deep link to package and changelog to README.md
+-   Updated dependencies [cfe9c13]
+    -   @sap-ux/abap-deploy-config-sub-generator@0.1.93
+    -   @sap-ux/cf-deploy-config-sub-generator@0.2.78
+    -   @sap-ux/deploy-config-generator-shared@0.1.51
+    -   @sap-ux/fiori-generator-shared@0.13.32
+    -   @sap-ux/odata-service-inquirer@2.11.1
+    -   @sap-ux/inquirer-common@0.9.1
+    -   @sap-ux/project-access@1.32.8
+    -   @sap-ux/ui5-config@0.29.9
+    -   @sap-ux/btp-utils@1.1.5
+    -   @sap-ux/store@1.3.3
+
+## 0.4.42
+
+### Patch Changes
+
+-   Updated dependencies [10a3b57]
+    -   @sap-ux/odata-service-inquirer@2.11.0
+
+## 0.4.41
+
+### Patch Changes
+
+-   Updated dependencies [4ddcff3]
+    -   @sap-ux/odata-service-inquirer@2.10.0
+    -   @sap-ux/inquirer-common@0.9.0
+    -   @sap-ux/cf-deploy-config-sub-generator@0.2.77
+    -   @sap-ux/deploy-config-generator-shared@0.1.50
+    -   @sap-ux/abap-deploy-config-sub-generator@0.1.92
+
 ## 0.4.40
 
 ### Patch Changes